removed commented out code
