+++ src/main/java/frc/robot/subsystems/elevator/ElevatorIOSim.java
@@ -74,26 +74,6 @@ public class ElevatorIOSim extends ElevatorIOTalonFX implements SimMechanism {
     elevator.setLength(physicsSim.getPositionMeters());
   }
 
-  // @Override
-  // public void setElevatorVelocity(double velocity) {
-  //   elevatorMotor1.setControl(
-  //       elevator1VelocityRequest.withVelocity(velocity /
-  // ElevatorSimConstants.METERS_PER_ROTATION));
-  //   elevatorMotor2.setControl(
-  //       elevator2VelocityRequest.withVelocity(velocity /
-  // ElevatorSimConstants.METERS_PER_ROTATION));
-  // }
-
-  // @Override
-  // public void setElevatorPosition(double position, int slot) {
-  //   elevatorMotor1.setControl(
-  //       elevator1PositionRequest.withPosition(position /
-  // ElevatorSimConstants.METERS_PER_ROTATION));
-  //   elevatorMotor2.setControl(
-  //       elevator2PositionRequest.withPosition(position /
-  // ElevatorSimConstants.METERS_PER_ROTATION));
-  // }
-
   @Override
   public double[] getCurrents() {
     return new double[] {physicsSim.getCurrentDrawAmps()};
@@ -102,9 +82,4 @@ public class ElevatorIOSim extends ElevatorIOTalonFX implements SimMechanism {
   public LoggedMechanismLigament2d getLigamentEnd() {
     return elevator;
   }
-
-  // @Override
-  // public void elevatorOpenLoop(Voltage voltage) {
-  //   elevatorMotor1.setVoltage(voltage.magnitude());
-  // }
 }
